- Moved all items under Plugin namespace into Plugin folder
