+++ MarkerMetro.Unity.WinLegacyUnity.Tests/MarkerMetro.Unity.WinLegacyUnity.Tests.csproj
@@ -50,19 +50,19 @@
     </Otherwise>
   </Choose>
   <ItemGroup>
-    <Compile Include="System\Collections\CollectionsExtensionsTests.cs" />
-    <Compile Include="System\Collections\ArrayListTests.cs" />
-    <Compile Include="System\Collections\HashtableTests.cs" />
-    <Compile Include="System\Collections\Specialized\OrderedDictionaryTests.cs" />
-    <Compile Include="System\Collections\Specialized\NameValueCollectionTests.cs" />
-    <Compile Include="System\IO\FileTests.cs" />
-    <Compile Include="System\IO\DirectoryTests.cs" />
-    <Compile Include="System\IO\StreamReaderExtensionsTests.cs" />
-    <Compile Include="System\Net\Sockets\TcpClientTests.cs" />
+    <Compile Include="Plugin\Collections\CollectionsExtensionsTests.cs" />
+    <Compile Include="Plugin\Collections\ArrayListTests.cs" />
+    <Compile Include="Plugin\Collections\HashtableTests.cs" />
+    <Compile Include="Plugin\Collections\Specialized\OrderedDictionaryTests.cs" />
+    <Compile Include="Plugin\Collections\Specialized\NameValueCollectionTests.cs" />
+    <Compile Include="Plugin\IO\FileTests.cs" />
+    <Compile Include="Plugin\IO\DirectoryTests.cs" />
+    <Compile Include="Plugin\IO\StreamReaderExtensionsTests.cs" />
+    <Compile Include="Plugin\Net\Sockets\TcpClientTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="System\Net\WebRequestTests.cs" />
+    <Compile Include="Plugin\Net\WebRequestTests.cs" />
     <Compile Include="System\Reflection\ReflectionTests.cs" />
-    <Compile Include="System\Collections\SortedListTests.cs" />
+    <Compile Include="Plugin\Collections\SortedListTests.cs" />
     <Compile Include="System\Security\Cryptography\CryptographyTests.cs" />
     <Compile Include="System\Threading\ThreadingTests.cs" />
     <Compile Include="System\Timers\TimerTests.cs" />
