+++ MarkerMetro.Unity.WinLegacyUnity.Tests/Plugin/IO/StreamReaderExtensionsTests.cs
