+++ MarkerMetro.Unity.WinLegacyUnity/MarkerMetro.Unity.WinLegacyUnity.csproj
@@ -44,41 +44,41 @@
     <Compile Include="Dispatcher.cs" />
     <Compile Include="System\AppDomain.cs" />
     <Compile Include="System\ApplicationException.cs" />
-    <Compile Include="System\Collections\ArrayList.cs" />
-    <Compile Include="System\Collections\CollectionsExtensions.cs" />
-    <Compile Include="System\Collections\Hashtable.cs" />
-    <Compile Include="System\Collections\SortedList.cs" />
-    <Compile Include="System\Collections\Specialized\IOrderedDictionary.cs" />
-    <Compile Include="System\Collections\Specialized\NameValueCollection.cs" />
-    <Compile Include="System\Collections\Specialized\OrderedDictionary.cs" />
+    <Compile Include="Plugin\Collections\ArrayList.cs" />
+    <Compile Include="Plugin\Collections\CollectionsExtensions.cs" />
+    <Compile Include="Plugin\Collections\Hashtable.cs" />
+    <Compile Include="Plugin\Collections\SortedList.cs" />
+    <Compile Include="Plugin\Collections\Specialized\IOrderedDictionary.cs" />
+    <Compile Include="Plugin\Collections\Specialized\NameValueCollection.cs" />
+    <Compile Include="Plugin\Collections\Specialized\OrderedDictionary.cs" />
     <Compile Include="System\Security\Cryptography\DESCryptography.cs" />
     <Compile Include="System\Security\Cryptography\EncryptionProvider.cs" />
     <Compile Include="System\Security\Cryptography\MD5CryptoServiceProvider.cs" />
     <Compile Include="System\Security\Cryptography\SHA1.cs" />
     <Compile Include="System\Diagnostics\StackFrame.cs" />
-    <Compile Include="System\IO\BinaryReaderExtensions.cs" />
-    <Compile Include="System\IO\BinaryWriterExtensions.cs" />
-    <Compile Include="System\IO\FileAccess.cs" />
-    <Compile Include="System\IO\DirectoryInfo.cs" />
-    <Compile Include="System\IO\FileInfo.cs" />
-    <Compile Include="System\IO\FileMode.cs" />
-    <Compile Include="System\IO\FileShare.cs" />
-    <Compile Include="System\IO\FileStream.cs" />
-    <Compile Include="System\IO\MemoryStreamExtensions.cs" />
-    <Compile Include="System\IO\ReadWriteStream.cs" />
-    <Compile Include="System\IO\SearchOption.cs" />
-    <Compile Include="System\IO\StreamExtensions.cs" />
-    <Compile Include="System\IO\StreamReader.cs" />
-    <Compile Include="System\IO\StreamReaderExtensions.cs" />
-    <Compile Include="System\IO\StreamWriter.cs" />
-    <Compile Include="System\IO\TextReaderExtensions.cs" />
-    <Compile Include="System\IO\Directory.cs" />
-    <Compile Include="System\IO\File.cs" />
+    <Compile Include="Plugin\IO\BinaryReaderExtensions.cs" />
+    <Compile Include="Plugin\IO\BinaryWriterExtensions.cs" />
+    <Compile Include="Plugin\IO\FileAccess.cs" />
+    <Compile Include="Plugin\IO\DirectoryInfo.cs" />
+    <Compile Include="Plugin\IO\FileInfo.cs" />
+    <Compile Include="Plugin\IO\FileMode.cs" />
+    <Compile Include="Plugin\IO\FileShare.cs" />
+    <Compile Include="Plugin\IO\FileStream.cs" />
+    <Compile Include="Plugin\IO\MemoryStreamExtensions.cs" />
+    <Compile Include="Plugin\IO\ReadWriteStream.cs" />
+    <Compile Include="Plugin\IO\SearchOption.cs" />
+    <Compile Include="Plugin\IO\StreamExtensions.cs" />
+    <Compile Include="Plugin\IO\StreamReader.cs" />
+    <Compile Include="Plugin\IO\StreamReaderExtensions.cs" />
+    <Compile Include="Plugin\IO\StreamWriter.cs" />
+    <Compile Include="Plugin\IO\TextReaderExtensions.cs" />
+    <Compile Include="Plugin\IO\Directory.cs" />
+    <Compile Include="Plugin\IO\File.cs" />
     <Compile Include="System\Math.cs" />
     <Compile Include="System\MissingExtensions.cs" />
     <Compile Include="System\Net\Sockets\TcpClient.cs" />
-    <Compile Include="System\Net\WebHeaderCollection.cs" />
-    <Compile Include="System\Net\WebRequest.cs" />
+    <Compile Include="Plugin\Net\WebHeaderCollection.cs" />
+    <Compile Include="Plugin\Net\WebRequest.cs" />
     <Compile Include="System\Reflection\ReflectionExtensions.cs" />
     <Compile Include="System\Reflection\TargetException.cs" />
     <Compile Include="System\Runtime\Remoting\Messaging\AsyncResult.cs" />
