+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/Collections/Specialized/NameValueCollection.cs
