+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/BinaryReaderExtensions.cs
