+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/BinaryWriterExtensions.cs
