+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/FileAccess.cs
