+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/FileMode.cs
