+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/FileShare.cs
