+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/MemoryStreamExtensions.cs
