+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/StreamExtensions.cs
