+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/StreamReader.cs
