+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/StreamReaderExtensions.cs
