+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/IO/StreamWriter.cs
