+++ MarkerMetro.Unity.WinLegacyUnity/Plugin/Net/WebRequest.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace MarkerMetro.Unity.WinLegacy.Net
+namespace MarkerMetro.Unity.WinLegacy.Plugin.Net
 {
     /// <summary>
     /// MSDN reference: http://msdn.microsoft.com/en-us/library/system.net.webrequest.aspx.
