--- MarkerMetro.Unity.WinLegacyUnity/System/IO/TextReaderExtensions.cs
@@ -1,12 +1,0 @@
-﻿using System.IO;
-
-namespace MarkerMetro.Unity.WinLegacy.Plugin.IO
-{
-    public static class TextReaderExtensions
-    {
-        public static void Close(this TextReader reader)
-        {
-            reader.Dispose();
-        }
-    }
-}
